Fix Hacker News seems to be down error while browsing.
Post the connection failure error with the didRaiseConnectionFailure
notification so that observers can figure out what went wrong.

The list view controller can then inspect this error and figure out
what to do. When quickly browsing through articles in the list some
requests appear to be cancelled and the connection fails with a
NSURLErrorCancelled error. This is not big deal and we don’t need to
show the failure message if it occurs.

Also add the notification and dictionary key as a string constant since
it makes the API more discoverable and less prone to be broken by
changing a string somewhere but forgetting to change it somewhere else
(this should really be done for every string in the project).

Also make sure that the notification observations are teared down when
the instance goes away (same, this should be cleaned across the whole
project).